+++ pages/complaint/index.jsx
@@ -9,7 +9,7 @@ import "swiper/css";
 import "swiper/css/autoplay";
 import "swiper/css/pagination";
 import { Autoplay, Pagination } from "swiper/modules";
-import CardModalDetail from "@/components/CardModalDetail";
+import CardModalDetail from "@/components/complaints/CardModalDetail";
 import { Clock, CheckCircle2, FileText, UserCheck, MapPin, Calendar } from "lucide-react";
 
 // Step indicator component
